Add files via upload
This project done in ZCIENTIA LAB. In this project using HTML, CSS, JavaScript, MySQL and Advance Java. This is the web page program. First iam design figma design framework after that worked on. As a Employee how to login a website in Company. I have mention every thing like Employee details and Account, Family . As a Employee based on the company fill the form in a company. This project very helpful for me and i have gained so many language. In a company how much required as know i have done. or in this project five language program code added.